+++ framework/alpas.gradle
@@ -21,7 +21,8 @@ dependencies {
     compile "org.eclipse.jetty:jetty-server:${jettyVersion}"
     compile "com.zaxxer:HikariCP:3.4.1"
     compile "org.atteo:evo-inflector:1.2.2"
-    compile "io.pebbletemplates:pebble:3.1.2"
+//    compile "io.pebbletemplates:pebble:3.1.2"
+    compile "com.github.PebbleTemplates:pebble:7e536a8894"
     compile "com.github.ajalt:clikt:2.3.0"
     compile "com.github.ajalt:mordant:1.2.1"
     compile "io.github.microutils:kotlin-logging:1.7.8"
